refactored ticket creation #SoClean
