+++ src/utils/graphics/center-position.ts
@@ -1,0 +1,6 @@
+export default function centerPosition(
+  containerSize: number,
+  elementSize: number
+) {
+  return Math.floor(containerSize / 2 - elementSize / 2);
+}
